Ironed out possible warnings in a object and framework code

+++ FrameLib_Dependencies/WindowFunctions.hpp
@@ -70,69 +70,69 @@ template <class T> void window_triangle(T window, uint32_t windowSize, uint32_t 
 {
     // FIX - this is approximate for now...
     
-    for (long i = 0; i < (windowSize >> 1) && i < generateSize; i++)
+    for (uint32_t i = 0; i < (windowSize >> 1) && i < generateSize; i++)
         window[i] = (double) i / (double) (generateSize >> 1);
-    for (long i = (windowSize >> 1); i < generateSize; i++)
+    for (uint32_t i = (windowSize >> 1); i < generateSize; i++)
         window[i] = (double) (((double) generateSize - 1) - (double) i) / (double) (generateSize >> 1);
 }
 
 template <class T> void window_hann(T window, uint32_t windowSize, uint32_t generateSize)
 {
-	for (long i = 0; i < generateSize; i++)
+	for (uint32_t i = 0; i < generateSize; i++)
 		window[i] = 0.5 - (0.5 * cos(WINDOW_TWOPI * normalise(i, windowSize)));
 }
 
 template <class T> void window_cosine(T window, uint32_t windowSize, uint32_t generateSize)
 {
-    for (long i = 0; i < generateSize; i++)
+    for (uint32_t i = 0; i < generateSize; i++)
         window[i] = sin(WINDOW_PI * normalise(i, windowSize));
 }
 
 template <class T> void window_hamming(T window, uint32_t windowSize, uint32_t generateSize)
 {
-	for (long i = 0; i < generateSize; i++)
+	for (uint32_t i = 0; i < generateSize; i++)
 		window[i] = 0.54347826 - (0.45652174 * cos(WINDOW_TWOPI * normalise(i, windowSize)));
 }
 
 template <class T> void window_blackman(T window, uint32_t windowSize, uint32_t generateSize)
 {
-	for (long i = 0; i < generateSize; i++)
+	for (uint32_t i = 0; i < generateSize; i++)
 		window[i] = 0.42659071 - (0.49656062 * cos(WINDOW_TWOPI * normalise(i, windowSize))) + (0.07684867 * cos(WINDOW_FOURPI * normalise(i, windowSize)));
 }
 
 template <class T> void window_blackman_62(T window, uint32_t windowSize, uint32_t generateSize)
 {
-	for (long i = 0; i < generateSize; i++)
+	for (uint32_t i = 0; i < generateSize; i++)
 		window[i] = (0.44859f - 0.49364f * cos(WINDOW_TWOPI * normalise(i, windowSize)) + 0.05677f * cos(WINDOW_FOURPI * normalise(i, windowSize)));
 }
 
 template <class T> void window_blackman_70(T window, uint32_t windowSize, uint32_t generateSize)
 {
-	for (long i = 0; i < generateSize; i++)
+	for (uint32_t i = 0; i < generateSize; i++)
 		window[i] = (0.42323f - 0.49755f * cos(WINDOW_TWOPI * normalise(i, windowSize)) + 0.07922f * cos(WINDOW_FOURPI * normalise(i, windowSize)));
 }
 
 template <class T> void window_blackman_74(T window, uint32_t windowSize, uint32_t generateSize)
 {
-	for (long i = 0; i < generateSize; i++)
+	for (uint32_t i = 0; i < generateSize; i++)
 		window[i] = (0.402217f - 0.49703f * cos(WINDOW_TWOPI * normalise(i, windowSize)) + 0.09892f * cos(WINDOW_FOURPI * normalise(i, windowSize)) - 0.00188 * cos(WINDOW_THREEPI * normalise(i, windowSize)));
 }
 
 template <class T> void window_blackman_92(T window, uint32_t windowSize, uint32_t generateSize)
 {
-	for (long i = 0; i < generateSize; i++)
+	for (uint32_t i = 0; i < generateSize; i++)
 		window[i] = (0.35875f - 0.48829f * cos(WINDOW_TWOPI * normalise(i, windowSize)) + 0.14128f * cos(WINDOW_FOURPI * normalise(i, windowSize)) - 0.01168 * cos(WINDOW_THREEPI * normalise(i, windowSize)));
 }
 
 template <class T> void window_blackman_harris(T window, uint32_t windowSize, uint32_t generateSize)
 {
-	for (long i = 0; i < generateSize; i++)
+	for (uint32_t i = 0; i < generateSize; i++)
 		window[i] = 0.35875 - (0.48829 * cos(WINDOW_TWOPI * normalise(i, windowSize))) + (0.14128 * cos(WINDOW_FOURPI * normalise(i, windowSize))) - (0.01168 * cos(WINDOW_SIXPI * normalise(i, windowSize)));
 }
 
 template <class T> void window_flat_top(T window, uint32_t windowSize, uint32_t generateSize)
 {
-	for (long i = 0; i < generateSize; i++)
+	for (uint32_t i = 0; i < generateSize; i++)
 		window[i] = 0.2810639 - (0.5208972 * cos(WINDOW_TWOPI * normalise(i, windowSize))) + (0.1980399 * cos(WINDOW_FOURPI * normalise(i, windowSize)));
 }
